Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Your code seems to just match Shay's algorithm though, unless I missed something. Do you remember what the further optimization was?
> 
> I had written a note to Shay that said I was able to eliminate some from_mont/to_mont conversion(s) from what he'd suggested in his paper "when blinding is used." I also remember thinking that in the end I'd made it so RSA blinding was close to free, except for the periodic random value generation; i.e. the blinding multiplications replaced some of the to_mont/from_mont conversions. (*ring* always does blinding.)

Oh, interesting. To be honest, I'm not entirely sure where you are and aren't in Montgomery form, so I could have missed that. Do you take advantage of your primes being of the same size so pq's R is the square of p and q's R, or something to that effect? (Something like that is probably more trouble than is worth it for us, at least until we manage to get rid of all stupidly-sized keys altogether. Which would be lovely, but I haven't tried yet.)

Or perhaps you meant elsewhere? I suppose I should study your code more closely sometime it's not Friday evening before a holiday. :-)

> > I assume you reduce c mod q^2 before doing elem_exp_consttime so that it's bounded by M^2 before doing a Montgomery reduction? I believe it works for inputs up to M*R. So as long as p and q use the same R, that shouldn't be necessary, right? Were tighter bounds just easier to encode (getting them into the typesystem like that is neat! I wish we could write this in a less unhelpful language...) and the subtraction too insignificant to worry about?
> 
> I just wasn't confident in my understanding of the math and also I wasn't confident that every code path could handle partially-reduced inputs, so I reduced them myself. Also I felt I could do the logic in Coq easier if I could assume inputs are already reduced (I barely can do anything in Coq). The subtraction seemed like a small price to pay considering all that.

Fair enough. Yeah, Andres has been teaching me how to use Coq and even so I'm still wrestling with the basics. :-) I think I can read the less insane parts of fiat-crypto okay now, but producing code still takes forever.

Transcribing all the places where we're using the slightly looser bounds would definitely be a thing to do.

> > I see you got rid of the RSAZ code, which avoids the dilemma around taking the BN_to_montgomery out of BN_mod_exp_mont_consttime.
> 
> IIRC, the thing that pushed me over the top was a conversation with Andy or Vlad or Shay where it was asserted that the RSAZ code is only the most optimal code for Haswell. I just didn't (don't) think it was a good use of time to maintain that code considering that and other factors.

Right, that's because on < Haswell the code can't run and on > Haswell, the ADX code is used, which we (and *ring* it appears) currently disable. :-P (I also suspect we do want the Haswell optimizations.)

> Also, in your comment you suggest we should always normalize the private key so that p > q. That is what *ring* requires of its inputs currently. However, I think it's also fine to have two code paths: one for p > q, and one for p < q. This would have a timing leak indicating which of those two values is greater, but I don't think that's significant. I seem to remember thinking that calculating ipmq was something I wanted to avoid, though I've forgotten why.

Sure, two codepaths where we swap the letters and compute ipmq rather than using the original iqmp is more-or-less the same as normalizing the keys? I guess the separate codepaths doesn't carry the risk of callers getting unhappy that their p and q are silently swapped, so that's useful in terms of needless breakage risk. (The curse of having existing consumers...)

Agreed that the timing leak is insignificant.

Patch-set: 2