Update patch set 2
Patch Set 2:

> To be honest, I'm not entirely sure where you are and aren't in Montgomery form, so I could have missed that.

We have different levels of Montgomery encoding and the Montgomery encoding level is encoded in the types. A value of type `One<P, Unencoded>` is the number 1 in field Zp, unencoded. A `Elem<P, RRR>` is some value in that field triple-Montgomery-encoded, and `Elem<P, RInverse>` is a value in that field that has been over-Montgomery-decoded.

https://github.com/briansmith/ring/blob/8411c0098ac1f8b4b96d0766a5057aae5e78c42a/src/arithmetic/montgomery.rs#L15-L38

> Do you take advantage of your primes being of the same size so pq's R is the square of p and q's R, or something to that effect? (Something like that is probably more trouble than is worth it for us, at least until we manage to get rid of all stupidly-sized keys altogether. Which would be lovely, but I haven't tried yet.)

Not directly. But we do take advantage of the fact that elements of Zp and Zq are also elements of Zn.

> Sure, two codepaths where we swap the letters and compute ipmq rather than using the original iqmp is more-or-less the same as normalizing the keys?

Probably we're talking past each other a bit. Happy to talk to you about it in the new year.

Patch-set: 2