Update patch set 1
Patch Set 1:

Thanks for the heads up! Though, annoyingly, we can't quite run the bots standalone with the r16 NDK just yet. (This doesn't affect Chromium or anyone else. This DEPS file is only used in BoringSSL's standalone builders.)

The runtime linker prints something to stderr which our tests can't run. I'm sure it's easy to fix, and I'm eager to use the NDK's CMake file. We're just all in Zurich for a conference right now, so we haven't had a chance to poke at it yet.
https://build.chromium.org/p/client.boringssl/builders/android_aarch64/builds/1628/steps/ssl%20tests/logs/stdio

In the meantime, we're actually still pinned to a pre-r16-NDK android_tools.git, so this shouldn't block anything on your end. Even when you remove the DEPS entry from android_tools.git master, the old revision will still recursedeps fine. (Let me know if that's not the case.)

We'll also need to update the recipes to pick up the NDK at the new location, but that's also no big deal.

How does this plan sound?

0. (You all continue with your transition plan without worrying about BoringSSL.)
1. We sort out the runtime linker error and reland https://boringssl-review.googlesource.com/c/boringssl/+/24644, so android_tools.git is caught up to r16.
2. Land this CL. We now have two identical copies of the NDK.
3. We update the boringssl.py recipe to grab the NDK out of util/bot/android_ndk
4. We update android_tools.git to an eventual later revision that I assume will remove the android_tools -> android_ndk DEPS entry and remove the recursedeps entry.

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>