Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Could you elaborate on why you need to build BoringSSL in these environments? (We usually want to have builder testing for any platform that we support.)
> 
> I came across various older embedded platforms (home routers, NAS boxes, ...) which are often using ARMv5 or even PPC SoCs, with corresponding old tool chains (e.g. gcc version 4.1 to 4.6).
> 
> E.g. one well known project depending on boringssl is mono. And plenty of other projects depend on mono. Mono supports various really old platforms perfectly fine. But if it's compiled with gcc < 4.8 it will disable the use of boringssl and fallback to their own legacy SSL/TLS implementation which only supports TLS 1.0. This fallback mechanism then causes all kind of trouble for the various .NET projects with simply expect a modern TLS stack.

BoringSSL is open-source, anyone can use it, but we generally don't take patches to accommodate them. Additionally, having a gray area for supported compilers is bad because we'll probably keep on breaking them, you'll keep on being frustrated that we keep on breaking them, and we'll keep on dealing with patches and #ifdefs to keep support for things that we only sort of support.

When it comes to making these decisions, we're trying to align with Abseil's rules (https://abseil.io/about/compatibility), which puts the end date on GCC 4.8 compatibility at 2018-03-23. But that is in the future, albeit only just. Thus, technically, I think we still support GCC 4.7 and so should take this patch. (However, come April, that'll no longer be the case and we may even remove this case along with any other EOLed support code at that point.)

Please check that version 2 still meets your needs. The only important change is that I switched '<=' to '<', otherwise I think that case would trigger for 4.8.0, which shouldn't need it, right?

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>