Add initial support for non-minimal BIGNUMs.
OpenSSL BIGNUMs are unsuitable for cryptography. It is a calculator
BIGNUM implementation. This is a problem seeing as we are a crypto
library. We'll need to either excise BIGNUM from RSA and EC or somehow
fix BIGNUM. EC is well-suited for the former (EC_SCALAR and later work
on field elements), but RSA's BIGNUMs are more transparent.  Teaching
BIGNUM to handle non-minimal word widths is probably simpler.

The main constraint is BIGNUM's large API surface. Some consumers use it
as a calculator. The API also gives access to RSA keys are BIGNUMs.
This must work with non-minimal BIGNUMs. As a first step, fix
comparison-type functions that were assuming minimal BIGNUMs. I've also
added bn_resize_words, but it is testing-only until the rest of the
library is fixed.

With non-minimal BIGNUMs allowed, the question is what widths should
they all have. I'm thinking we make width management implicit based on
what functions where used. The alternative would be to mimic EC_SCALAR
and add types like RSA_MOD_P, RSA_MOD_N, RSA_SCALAR_P, RSA_SCALAR_N...
that's a bit unwieldy in C. The public RSA structures are a nuisance
(https://boringssl-review.googlesource.com/c/boringssl/+/24445 is silly)
and RSA is a bit big to stack-allocate it all. I suspect, were we to do
that, it ought to be built on top of non-minimal BIGNUMs rather than
parallel to?

Non-miminal BIGNUMs means bn->top is no longer a property of the value,
but a loose upper bound that is carried around. It does not affect the
results of computation, only performance and secrecy. This is a
departure from the original meaning and compiler help in auditing
everything is nice, so the final change in this series will rename
bn->top to bn->width. Thus these new functions are named per "width",
not "top".

Output widths will be handled on a case-by-case basis:

- Calculator functions like BN_exp will just do what is simplest and
  probably continue to treat things as minimal.

- Trivially-bounded functions like modular arithmetic will set the
  output width to the modulus width. This covers the majority of crypto
  code and is easy to make constant-time.

- Functions like BN_add, BN_mul, or BN_sqr which grow but are
  occasionally used in crypto[*] are a little fuzzier. Constant-time
  dictates that BN_mul sets r->top = a->top + b->top, etc., but a
  calculator repeatedly multiplying by BN_value_one would then run out of
  memory. Those we'll split into a private BN_mul_fixed for crypto,
  leaving BN_mul for calculators. BN_mul is just BN_mul_fixed followed
  by bn_correct_top.

Note this scheme makes it non-obvious what widths are for our
intermediate values. Code that assumes a particular output width of
BN_mod_mul_montgomery could work, but it's messy. There's three notions
of correctness here:

1. Do I compute the right answer?

2. Do I handle secret data in constant time?

3. Does my memory usage not balloon absurdly?

The simplest approach is to write code in a width-agnostic way and rely
on functions to accept all widths. This means staying in the BIGNUM
abstraction where possible rather than dipping into "words". If all
BIGNUM functions handle all input widths, (1) is resolved. (2) isn't any
different than usual. Don't call non-constant-time functions on secret
data. (3) is a loose bound and, in most cases, will happen for free.
If pessimistic bounds are ever too loose, we add explicit
bn_resize_words calls.

Where we do dip into "words" (e.g. the BIGNUM functions themselves),
things are messier. All input widths must be tolerated. For calculator
logic, bn_minimal_width returns the original bn->top semantics and can
be freely called. For constant-time logic, we must handle more cases at
runtime. Although BN_mod_add_quick can assume reduced inputs, |a| could
be wider or narrower than |m|. But any words in |a| beyond |m| are known
to be zero, so helpers like bn_resize_words can smooth over oversized
inputs. Undersized immutable inputs need manual care, but they were
already possible with minimal-width BIGNUMs.

[*] A lot of these functions are also not constant-time for other
reasons, but that will be fixed separately.

Bug: 232
Change-Id: Ide2258ae8c09a9a41bb71d6777908d1c27917069