Fix up BN_MONT_CTX_set with non-minimal values.
I suspect things would work if we tried to size R based on the
non-minimal width, but the API promise is that widths do not affect the
answers, thus it should be minimal. Either way we don't have immediate
need[*] for a non-minimal one.

With the top explicitly made minimal, computing |lgBigR| is also a
little simpler.

This CL also abstracts out the < R check in the RSA code, and implements
it in a width-agnostic way.

[*] Well, I could imagine doing the mismatched p and q cases for RSA
that way, but it may be easier to just try to reject such keys or deny
them CRT. But we'll deal with that if it comes up. An explicit parameter
may be better than implicitly using the width.

Bug: 232
Change-Id: I354643df30530db7866bb7820e34241d7614f3c2