Update patch set 1
Patch Set 1:

(2 comments)

> Patch Set 1:
> 
> (2 comments)
> 
> So I understand you mostly want this in transition, but I'm a little curious what actually wants this. I've always seen SSL_VERIFY_NONE as a historical wart that we just haven't had the time to confirm unnecessary. Are there consumers for whom SSL_VERIFY_NONE is actually meaningfully different from just making the callback blindly accept everything? The only difference is that SSL_get_verify_result() will show you an error, but if you were going to reject on that, you ought to have just made the callback implement that...

I went with this approach because we have a caller that wants (in some cases) only to collect the result of verification, while allowing the connection to succeed.  SSL_VERIFY_NONE expresses their intent well.

The caller also has a custom callback installed with SSL_CTX_set_verify().  They could modify their callback to notice when |ok| is false, and (if called with ok = 0) stash that information somewhere, and return 1.  But then, is it guaranteed that every call to the callback with ok = 0 is dispositive, with altchains enabled?  It gets hairy.

That said: if you want to ban SSL_VERIFY_NONE as the wart that it is, I can totally work with that.  But then I think we should explicitly reject it as an argument to SSL_CTX_set_custom_verify().

Patch-set: 1
Reviewer: Gerrit User 5645 <5645@66b711a4-aa9b-3dac-9062-916540830212>