Update patch set 2
Patch Set 2:

I wonder if this design went too far in saying the widths are non-existent. It's very attractive in terms of making it "obvious" that random incremental fixes work, or what the semantics of any particular function should be. It also avoids worrying about length-matches that C's typesystem is too poor to enforce for us.

But boy is it obnoxious to implement anything that needs to do a select. :-/ I'll ponder more on this later after dinner...

Patch-set: 2
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0