Make bn_mul_recursive constant-time.
I left the input length as int because the calling convention passes
these messy deltas around. This micro-optimization is almost certainly
pointless, but bn_sub_part_words is written in assembly, so I've left it
alone for now.

This leaves bn_mul_part_recursive, which is reachable for RSA keys which
are not a power of two in size.

Bug: 234
Change-Id: I72848d992637c0390cdd3c4f81cb919393b59eb8