Make primality testing mostly constant-time.
The extra details in Enhanced Rabin-Miller are only used in
RSA_check_key_fips, on the public RSA modulus, which the static linker
will drop in most of our consumers anyway. Implement normal Rabin-Miller
for RSA keygen and use Montgomery reduction so it runs in constant-time.

Note that we only need to avoid leaking information about the input if
it's a large prime. If the number ends up composite, or we find it in
our table of small primes, we can return immediately.

The leaks not addressed by this CL are:

- The difficulty of selecting |b| leaks information about |w|.
- The distribution of whether step 4.4 runs leaks information about w.
- We leak |a| (the largest power of two which divides w) everywhere.
- BN_mod_word in the trial division is not constant-time.

These will be resolved in follow-up changes.

This makes even RSA keygen a hair faster, now that we don't have the
unnecessary GCDs in the way. Montgomery reduction is also a good idea
when doing many modular multiplications. A median keygen time (out of 10
samples) of 0.753s goes to 0.715s, though this is not really a
statistically significant measurement.

Bug: 238
Change-Id: I0cf0ff22079732a0a3ababfe352bb4327e95b879