Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

For code within BoringSSL, we also have CBBFinishArray in ssl/internal.h which does approximately this with a different set of scopers. I noticed you used this in tests. Was the plan to use this in consuming code too, or just our tests?

I like bssl::Array because it carries its length with it (it's an owning version of bssl::Span), but in consumers it's basically strictly worse than std::vector and absl::FixedArray. You could imagine a function which makes a Span out of CBB_data/CBB_len, and then the caller does what they like with it, but the STL has not yet been recast in terms of spans and iterator pairs yet. :-/

(Also fine just adding this helper and not thinking too hard about it. Doesn't really matter.)

Patch-set: 3
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=+1