Blind the range check for finding a Rabin-Miller witness.
Part of Rabin-Miller requires selecting a random number from 2 to |w|-1.
This is done by picking an N-bit number and discarding out-of-range
values. To avoid leaking information, add |bn_rand_secret_range| which,
rather than discarding bad values, fixes them to be in range.  Though
not uniformly selected, these fixed values are still usable as potential
Rabin-Miller witnesses.

We keep counters of both how many total iterations we've run, and how
many iterations were selected uniformly. We iterate until we've both
performed at least |iterations| uniform iterations and at least
|BN_PRIME_CHECKS_BLINDED| total iterations. If |BN_PRIME_CHECKS_BLINDED|
is set large enough, it will be the limiting factor with high
probability and we won't leak information.

Rabin-Miller is already probabilistic, so we could reach the desired
confidence levels with this non-uniform distribution by simply
increasing the iteration count appropriately. However, to align with
FIPS 186-4, we use the more pessimal analysis. As a result, this
function is more complex and has more timing risk than is actually
necessary.

Note that because composite numbers may still be rejected in
variable-time by a discarded value, this blinding does not impact the
majority of candidate RSA factors in RSA key generation. Only the final
few primality tests that actually land on a prime do more work.  Thus
this does make the BNTest.PrimeChecking test take about 2x longer to run
on debug mode, but RSA key generation time is unaffected. A median of
0.251s becomes 0.257s.

Another, perhaps simpler, option here would have to run
bn_rand_range_words to the full 100 count, select an arbitrary
successful try, and declare failure of the entire keygen process (as we
do already) if all tries failed. I went with the option in this CL
because I happened to come up with it first, and because the failure
probability decreases much faster. Additionally, the option in this CL
does not affect composite numbers, while the alternate would. This gives
a smaller multiplier on our entropy draw. We also continue to use the
"wasted" work for stronger assurance on primality. FIPS' numbers are
remarkably low, considering the increase has negligible cost.

Thanks to Nathan Benjamin for helping me explore the failure rate as the
target count and blinding count change.

Now we're down to the rest of RSA keygen, which will require all the
operations we've traditionally just avoided in constant-time code!

Bug: 238
Change-Id: Id6406c3020f2585b86946eb17df64ac42f30ebab